Streamline field utility and entity utility
